atest.py -> test.py (-a runs AMBER)
